simpler search logic
