+++ src/hooks/useSuggestions.ts
@@ -1,0 +1,23 @@
+import { useState, useRef, useEffect } from 'react';
+
+export const useSuggestions = () => {
+  const [showSuggestions, setShowSuggestions] = useState(false);
+  const suggestionRef = useRef<HTMLDivElement>(null);
+
+  useEffect(() => {
+    const handleClickOutside = (event: MouseEvent) => {
+      if (suggestionRef.current && !suggestionRef.current.contains(event.target as Node)) {
+        setShowSuggestions(false);
+      }
+    };
+
+    document.addEventListener('mousedown', handleClickOutside);
+    return () => document.removeEventListener('mousedown', handleClickOutside);
+  }, []);
+
+  return {
+    showSuggestions,
+    setShowSuggestions,
+    suggestionRef,
+  };
+};
